pyproject.toml
[project]
name = "mcp-weather"
version = "1.1.0"
description = "Weather forecast tool for MCP with HTTP streaming support"
authors = []
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "python-dotenv",
    "aiohttp",
    "uvicorn",
    "mcp[cli]>=1.14.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "vulture>=2.0.0",
    "bandit>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_weather"]

[project.scripts]
mcp-weather = "mcp_weather.weather:mcp.run"
